+++ src/main/java/com/esaulpaugh/headlong/rlp/RLPItem.java
@@ -15,11 +15,11 @@
 */
 package com.esaulpaugh.headlong.rlp;
 
-import com.esaulpaugh.headlong.rlp.exception.DecodeException;
-import com.esaulpaugh.headlong.rlp.exception.RecoverableDecodeException;
-import com.esaulpaugh.headlong.rlp.exception.UnrecoverableDecodeException;
+import com.esaulpaugh.headlong.util.exception.DecodeException;
+import com.esaulpaugh.headlong.util.exception.RecoverableDecodeException;
+import com.esaulpaugh.headlong.util.exception.UnrecoverableDecodeException;
 import com.esaulpaugh.headlong.rlp.util.FloatingPoint;
-import com.esaulpaugh.headlong.rlp.util.Integers;
+import com.esaulpaugh.headlong.util.Integers;
 import com.esaulpaugh.headlong.rlp.util.Notation;
 import com.esaulpaugh.headlong.util.Strings;
