+++ src/main/java/com/esaulpaugh/headlong/rlp/RLPListIterator.java
@@ -15,7 +15,7 @@
 */
 package com.esaulpaugh.headlong.rlp;
 
-import com.esaulpaugh.headlong.rlp.exception.DecodeException;
+import com.esaulpaugh.headlong.util.exception.DecodeException;
 
 import java.util.NoSuchElementException;
