+++ src/main/java/com/esaulpaugh/headlong/util/exception/UnrecoverableDecodeException.java
@@ -13,10 +13,10 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.esaulpaugh.headlong.rlp.exception;
+package com.esaulpaugh.headlong.util.exception;
 
 /**
- * Indicates a failure to decode an RLP item that is unrecoverably malformed or exceeds the bounds of its parent item.
+ * Indicates a failure to decode an item that is unrecoverably malformed.
  */
 public final class UnrecoverableDecodeException extends DecodeException {
