+++ src/test/java/com/esaulpaugh/headlong/example/StudentTest.java
@@ -17,7 +17,7 @@ package com.esaulpaugh.headlong.example;
 
 import com.esaulpaugh.headlong.abi.Function;
 import com.esaulpaugh.headlong.abi.TupleType;
-import com.esaulpaugh.headlong.rlp.exception.DecodeException;
+import com.esaulpaugh.headlong.util.exception.DecodeException;
 import com.esaulpaugh.headlong.util.Strings;
 import org.junit.jupiter.api.Test;
