+++ src/test/java/com/esaulpaugh/headlong/rlp/Benchmark.java
@@ -17,7 +17,7 @@ package com.esaulpaugh.headlong.rlp;
 
 import com.esaulpaugh.headlong.example.Student;
 import com.esaulpaugh.headlong.example.StudentTest;
-import com.esaulpaugh.headlong.rlp.exception.DecodeException;
+import com.esaulpaugh.headlong.util.exception.DecodeException;
 import com.esaulpaugh.headlong.rlp.util.Notation;
 import com.esaulpaugh.headlong.util.Strings;
 import org.junit.jupiter.api.Disabled;
