+++ src/test/java/com/esaulpaugh/headlong/rlp/RLPJsonEncodeTest.java
@@ -17,7 +17,7 @@ package com.esaulpaugh.headlong.rlp;
 
 import com.esaulpaugh.headlong.TestUtils;
 import com.esaulpaugh.headlong.abi.util.JsonUtils;
-import com.esaulpaugh.headlong.rlp.util.Integers;
+import com.esaulpaugh.headlong.util.Integers;
 import com.esaulpaugh.headlong.util.FastHex;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
