+++ src/test/java/com/esaulpaugh/headlong/rlp/RLPStreamIteratorTest.java
@@ -16,8 +16,8 @@
 package com.esaulpaugh.headlong.rlp;
 
 import com.esaulpaugh.headlong.TestUtils;
-import com.esaulpaugh.headlong.rlp.exception.DecodeException;
-import com.esaulpaugh.headlong.rlp.exception.UnrecoverableDecodeException;
+import com.esaulpaugh.headlong.util.exception.DecodeException;
+import com.esaulpaugh.headlong.util.exception.UnrecoverableDecodeException;
 import com.esaulpaugh.headlong.util.FastHex;
 import com.esaulpaugh.headlong.util.Strings;
 import org.junit.jupiter.api.Test;
