+++ src/test/java/com/esaulpaugh/headlong/rlp/util/FloatingPointTest.java
@@ -1,7 +1,7 @@
 package com.esaulpaugh.headlong.rlp.util;
 
 import com.esaulpaugh.headlong.abi.MonteCarloTest;
-import com.esaulpaugh.headlong.rlp.exception.DecodeException;
+import com.esaulpaugh.headlong.util.exception.DecodeException;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
